# Merge in-memory indexes; use multiple maps