… internal/repository/master_index_test.go