What is the purpose of this change? What does it change?
--------------------------------------------------------

Merge in-memory indexes when they are finalized.
This increases performance especially for large repositories with many index files.
To avoid large spikes in memory consumption when maps are resized, this PR also introduces sharded maps to save the index entries. By using 256 shards one byte of key information can be removed which also reduces the memory required for large indexes a bit.

Was the change discussed in an issue or in the forum before?
------------------------------------------------------------

The idea comes from #2523.

Checklist
---------

- [x] I have read the [Contribution Guidelines](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#providing-patches)
- [x] I have enabled [maintainer edits for this PR](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/allowing-changes-to-a-pull-request-branch-created-from-a-fork)
- [ ] I have added tests for all changes in this PR
- [ ] I have added documentation for the changes (in the manual)
- [ ] There's a new file in `changelog/unreleased/` that describes the changes for our users (template [here](https://github.com/restic/restic/blob/master/changelog/TEMPLATE))
- [x] I have run `gofmt` on the code in all commits
- [x] All commit messages are formatted in the same style as [the other commits in the repo](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#git-commits)
- [x] I'm done, this Pull Request is ready for review