@aawsome 2020-06-19
> I'm a bit worried about the memory overhead this PR might have. When the index hash map has to be resized, Go doubles the size of the backing array and then slowly migrates the entries into the new hashmap. That way we might end up needing three times the memory than actually needed.

I didn't dive into the golang map implementation too much, but seems you are right, this might be an issue. I tried to add an benchmark `BenchmarkMasterIndexLookupMultipleIndexMerged`and this failed with an memory allocation error on my 8GB machine..

I changed the benchmark (a 200k pack file index should't occur in reality) to use 50 indexes with 20k packs. The benchmark results of this much more realistic setting show a huge speed improvement.

I think that we should try to find a replacement for the `blobs` map - as the implementation of the map causes this trouble - and leave the Index and MasterIndex logic as is.

However, as long as we risk a huge memory overhead, this should not be merged;  I'll change this PR to draft mode.