@greatroar 2020-06-19
A simple sharded hash table grows at the same rate as a regular one, but avoids the big temporary (one shard grows at a time). It can also prevent lock contention. E.g., (untested)

```go
type idxMap [32]struct {
	sync.Mutex
	m map[restic.BlobHandle]indexEntry
}

func (m *idxMap) get(h restic.BlobHandle) (indexEntry, bool) {
	i := binary.BigEndian.Uint64(h.ID[:8])
	m[i].Lock()
	defer m[i].Unlock()

	entry, ok := m[i].m[h]
	return entry, ok
}
```