@aawsome 2020-06-19
@greatroar I was thinking about something like this.
Maybe even use 256 shards - this would allow us to use `h.ID[1:]` in the map key and saves 1 additional byte per index entry :smile: