@aawsome 2020-06-19
Ok, as implementation of the sharded index was not too complex, I've added this to this PR.

A first benchmark showed that the speed improvement of the merged index should outweight the additional overhead introduced by the shards. Also slightly reduces the memory used by the index for large repositories (but introduces the cost of a constant overhead to manage the 256 shards)