@aawsome 2020-06-19
Here are the benchmark results of `internal/repository` (the BenchmarkMasterIndexLookupMultipleIndexReal* benchmarks are added with this PR)

Old:
```
BenchmarkPackerManager-4                           	       4	 272467956 ns/op	 193.42 MB/s	   91178 B/op	    1411 allocs/op
BenchmarkDecodeIndex-4                             	   66586	     17468 ns/op
BenchmarkIndexHasUnknown-4                         	21096319	        57.1 ns/op
BenchmarkIndexHasKnown-4                           	19841572	        61.1 ns/op
BenchmarkIndexAlloc-4                              	       2	 964482906 ns/op	785897508 B/op	  976686 allocs/op
BenchmarkMasterIndexLookupSingleIndex-4            	 7649384	       155 ns/op
BenchmarkMasterIndexLookupMultipleIndex-4          	 2455197	       474 ns/op
BenchmarkMasterIndexLookupSingleIndexUnknown-4     	14100042	        85.4 ns/op
BenchmarkMasterIndexLookupMultipleIndexUnknown-4   	 3046612	       396 ns/op
BenchmarkSaveAndEncrypt-4                          	 8935536	       129 ns/op	32436382.67 MB/s	       2 B/op	       0 allocs/op
BenchmarkLoadTree-4                                	    3351	    354630 ns/op
BenchmarkLoadBlob-4                                	     140	   8288340 ns/op	 120.65 MB/s
BenchmarkLoadAndDecrypt-4                          	     100	  10114474 ns/op	  98.87 MB/s
BenchmarkLoadIndex-4                               	      30	  41519666 ns/op
```

New:
```
BenchmarkPackerManager-4                                     	       4	 267857058 ns/op	 196.75 MB/s	   91174 B/op	    1411 allocs/op
BenchmarkDecodeIndex-4                                       	   22050	     52617 ns/op
BenchmarkIndexHasUnknown-4                                   	18344560	        65.8 ns/op
BenchmarkIndexHasKnown-4                                     	17026891	        70.6 ns/op
BenchmarkIndexAlloc-4                                        	       1	1099485149 ns/op	759550568 B/op	  982149 allocs/op
BenchmarkMasterIndexLookupSingleIndex-4                      	 5698429	       221 ns/op
BenchmarkMasterIndexLookupMultipleIndex-4                    	 2507014	       506 ns/op
BenchmarkMasterIndexLookupMultipleIndexMerged-4              	 7457487	       165 ns/op
BenchmarkMasterIndexLookupSingleIndexUnknown-4               	 7913674	       144 ns/op
BenchmarkMasterIndexLookupMultipleIndexUnknown-4             	 2839654	       422 ns/op
BenchmarkMasterIndexLookupMultipleIndexMergedUnknown-4       	12912753	        92.0 ns/op
BenchmarkMasterIndexLookupMultipleIndexReal-4                	  149528	      8000 ns/op
BenchmarkMasterIndexLookupMultipleIndexRealMerged-4          	 7547283	       160 ns/op
BenchmarkMasterIndexLookupMultipleIndexRealUnknown-4         	  151024	      8089 ns/op
BenchmarkMasterIndexLookupMultipleIndexRealMergedUnknown-4   	12840684	        91.8 ns/op
BenchmarkSaveAndEncrypt-4                                    	 8539178	       140 ns/op	30004544.64 MB/s	       2 B/op	       0 allocs/op
BenchmarkLoadTree-4                                          	    3098	    357358 ns/op
BenchmarkLoadBlob-4                                          	     141	   8457648 ns/op	 118.24 MB/s
BenchmarkLoadAndDecrypt-4                                    	     115	   9897292 ns/op	 101.04 MB/s
BenchmarkLoadIndex-4                                         	      31	  40329784 ns/op
```

So pure Lookup within `Index` seems to be about 9ns slower than before and also with MasterIndex there is some more overhead (about 60ns per lookup) 
However with multiple indexes in MasterIndex you see the benefit already in the 5 index files à 200k packs example - compare e.g. old BenchmarkMasterIndexLookupMultipleIndex-4 with new BenchmarkMasterIndexLookupMultipleIndexMerged-4. 
In the more real world scenario (100 index files à 10k pack files) you can see that the merged index also scales well with a lot of indexes being loaded and merged - as expected.

Memory usage seems to be reduced by about 3%