@greatroar 2020-06-19
You can get a nice overview of benchmark results from the [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat) tool:

```
$ git checkout $oldversion
$ go test -bench=. -count=10 | tee old
$ git checkout $newversion
$ go test -bench=. -count=10 | tee new
$ benchstat old new
```